requirements.txt
absl-py==0.7.0
apturl==0.5.2
astor==0.7.1
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
defer==1.0.6
devscripts===2.16.2ubuntu3
easydict==1.9
feedparser==5.1.3
future==0.17.1
gast==0.2.2
gin-config==0.1.2
google-pasta==0.1.1
grpcio==1.18.0
guacamole==0.9.2
gym==0.10.9
h5py==2.9.0
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Magic-file-extensions==0.2
Mako==1.0.3
Markdown==3.0.1
MarkupSafe==0.23
matplotlib==3.0.2
mock==2.0.0
numpy==1.16.0
oauthlib==1.0.3
onboard==1.2.0
opencv-python==4.0.0.21
padme==1.1.1
pbr==5.1.1
pexpect==4.0.1
Pillow==5.4.1
pkg-resources==0.0.0
plainbox==0.25
protobuf==3.6.1
ptyprocess==0.5
pyasn1==0.1.9
-e git+https://github.com/bulletphysics/bullet3@6a74f63604ceecd1db5c71036ffb0dbf17294579#egg=pybullet
pycups==1.9.73
pycurl==7.43.0
pyglet==1.3.2
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.5
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
PyYAML==3.13
reportlab==3.3.0
requests==2.9.1
scipy==1.2.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
ssh-import-id==5.5
system-service==0.3
tb-nightly==1.13.0a20190117
tensorflow-probability==0.5.0
termcolor==1.1.0
tf-agents==0.2.0rc2
tf-agents-nightly==0.2.0.dev20181204
tf-estimator-nightly==1.13.0.dev2019010910
tf-nightly-gpu==1.13.0.dev20190117
tfp-nightly==0.6.0.dev20190115
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
